# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop API - Swagger Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/admin/create": {
            "post": {
                "operationId": "adfd219fe5c92529a9cb535d258c2f80",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/admin/login": {
            "post": {
                "operationId": "5a40c8e5833e64b90cb28af5048e12c2",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/admin/logout": {
            "get": {
                "operationId": "e87c8bf7d33c4a933cf7174a5d36a270",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/admin/user-delete/{uuid}": {
            "get": {
                "operationId": "f46c06f2db6bd9f22566d7899c39056c",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/admin/user-edit/{uuid}": {
            "put": {
                "operationId": "d1333ebad88321674c564aa94807ee16",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/admin/user-listing": {
            "put": {
                "operationId": "24bba07f28468afce53457c8ba2409f6",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "uuid": {
                        "type": "string"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "email_verified_at": {
                        "type": "string"
                    },
                    "avatar": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "string"
                    },
                    "is_marketing": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "last_login_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}